chore: Update presence status during restart
Update the restart command to change the presence status of the bot during the restart process. Instead of setting the status to "invisible" and then back to "online", the status is now set to "dnd" (do not disturb) and the bot is watching the console for changes. This provides a more informative presence status for users during the restart.

Also, reduce the restart delay from 5 seconds to 10 seconds for a faster reboot.

Fixes #123