chore: Update server-invite command to generate invite link for the bot
Refactor the server-invite command to generate an invite link for the bot to join the user's server. This improves the functionality of the command by providing a convenient way for users to invite the bot. Additionally, handle errors that may occur during the invite link generation process.

Fixes #N/A