chore: Add devonly flag to presence command
Add the `devonly` flag to the `presence` command in order to restrict its usage to users with the "Bot Owner" role. This ensures that only authorized users can set the bot's presence. The `devonly` flag is set to `true` in the command options. This improves the security and control of the bot's presence feature.

Fixes #N/A